Don't change the print_vec implementation between `nm` runs
I got confused when you were comparing the two calls to `nm` because the
code changed the signature and implementation of the print_vec function,
but was only trying to illustrate the change that happens because of the
invocation of print_vec on only integers, not strings too. So I think
print_vec should stay exactly the same to make the point clearer.